fix/deleteMemberFromChannel controller and resolved requested changes
